requirements.txt
gradio==4.41.0
transformers==4.44.0
accelerate==0.33.0
pillow==10.4.0
numpy==2.0.1
regex==2024.7.24
textwrap3==0.9.2
os-sys==2.1.4
torch==2.4.0
python-dotenv==1.0.1